requirements.txt
black<=24.10.0
python-docx<=1.1.2
num2words<=0.5.13
pymorphy3<=2.0.2
pymorphy3-dicts-uk<=2.4.1.1.1663094765